requirements.txt
flask>=2.3.0
ultralytics>=8.0.0
pillow>=10.0.0
